refactor: partner 타입 정리
