+++ src/states/partner.ts
@@ -1,11 +1,5 @@
+import { Company, FinancialData } from "@/types/partner";
 import { atom } from "jotai";
-
-// 회사 관련
-export interface Company {
-  id: string;
-  name: string;
-  type: "payment" | "collection";
-}
 
 export const companiesAtom = atom<Company[]>([]);
 export const selectedCompanyAtom = atom<string>("");
@@ -13,23 +7,6 @@ export const selectedCompanyAtom = atom<string>("");
 // 년도 관련
 export const availableYearsAtom = atom<number[]>([]);
 export const selectedYearAtom = atom<number>(2025);
-
-// 재무 데이터
-export interface FinancialData {
-  isCarryover?: boolean;
-  id: string;
-  year: number;
-  month: string;
-  lamplePurchase: number | null;
-  lamplePayment: number | null;
-  lampleBalance: number | null;
-  gompyoPurchase: number | null;
-  gompyoPayment: number | null;
-  gompyoBalance: number | null;
-  totalPurchase: number | null;
-  totalPayment: number | null;
-  totalBalance: number | null;
-}
 
 export const financialDataAtom = atom<FinancialData[]>([]);
